requirements.txt
python-dotenv
python-telegram-bot==22.3
requests
flask
